+++ setup/clean.sh
@@ -13,6 +13,10 @@ if [ -d ./data ]; then
   sudo rm -fr data
 fi
 
+if [ -e setup_ngsi_go.sh ]; then
+  sudo rm -f setup_ngsi_go.sh
+fi
+
 sudo rm -fr ./config
 rm -fr ./.work
 rm -f ./.env