+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## FIWARE Big Bang v0.20.0-next
 
+-   ADD feature_request.md (#244)
 -   ADD test cases (#243)
 -   Fix fail while setting up Draco (#242)
 -   Update Postgres to 15 (#241)