+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## FIWARE Big Bang v0.20.0-next
+## FIWARE Big Bang v0.21.0 - 31 January, 2023
 
 -   ADD feature_request.md (#244)
 -   ADD test cases (#243)