+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## FIWARE Big Bang v0.21.0-next
+
 ## FIWARE Big Bang v0.21.0 - 31 January, 2023
 
 -   ADD feature_request.md (#244)